+++ jina/peapods/__init__.py
@@ -4,4 +4,4 @@ __license__ = "Apache-2.0"
 from .peas import BasePea as Pea
 from .pods import BasePod
 from .pods import Pod
-from .compoundpod import CompoundPod
+from .pods.compoundpod import CompoundPod
